+++ web/database/migrations/2024_10_11_010855_create_msmes_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->integer('price_min');
             $table->integer('price_max');
             $table->json('schedules');
-            $table->double('rate');
+            $table->float('rate');
             $table->boolean('is_recomended');
             $table->timestamps();
         });
